refactor(anthropic): split monolithic client.go into focused files
client.go (790→408 lines, -48%): decompose into single-responsibility
files within the same package anthropic:

- messages.go (164 lines): message conversion pipeline —
  toAnthropicMessages, convertToAnthropicBlocks, partToContentBlock,
  appendOrMergeMessage, appendSystemContent, marshalResponse
- metrics.go (184 lines): response parsing & token accounting —
  fromAnthropicResponse, extractContent, buildMetrics,
  resolveTrafficType, checkTruncation
- tools.go (69 lines): tool schema conversion —
  toAnthropicTools, toAnthropicSchema
- client.go (408 lines): struct, constructor, options, SendChat,
  HTTP plumbing, types, utilities

All methods remain on *client receiver. No behavioral changes.
go test ./... passes all 65 packages including integration/e2e.
