+++ internal/infrastructure/llm/anthropic/client.go
@@ -303,377 +303,6 @@ func (c *client) SendChat(ctx context.Context, history []*llm.Content, toolDecls
 	return c.fromAnthropicResponse(&msgResp, totalDuration.Seconds())
 }
 
-func (c *client) toAnthropicMessages(history []*llm.Content) (string, []message, error) {
-	system := c.persona
-	messages := make([]message, 0, len(history))
-
-	for _, h := range history {
-		if h.Role == "system" {
-			system = c.appendSystemContent(system, h)
-			continue
-		}
-
-		role, blocks, err := c.convertToAnthropicBlocks(h)
-		if err != nil {
-			return "", nil, err
-		}
-		if len(blocks) == 0 {
-			continue
-		}
-
-		messages = c.appendOrMergeMessage(messages, role, blocks)
-	}
-
-	// Apply cache breakpoint to the last message in history to enable conversation history caching
-	if len(messages) > 0 {
-		lastMsg := &messages[len(messages)-1]
-		if len(lastMsg.Content) > 0 {
-			lastMsg.Content[len(lastMsg.Content)-1].CacheControl = &cacheControl{
-				Type: "ephemeral",
-			}
-		}
-	}
-
-	return system, messages, nil
-}
-
-func (c *client) appendSystemContent(currentSystem string, h *llm.Content) string {
-	parts := make([]string, 0, len(h.Parts))
-	for _, p := range h.Parts {
-		if p.Text != "" {
-			parts = append(parts, p.Text)
-		}
-	}
-	newContent := strings.Join(parts, "\n")
-	if currentSystem != "" {
-		return currentSystem + "\n\n" + newContent
-	}
-	return newContent
-}
-
-func (c *client) convertToAnthropicBlocks(h *llm.Content) (string, []contentBlock, error) {
-	role := h.Role
-	switch role {
-	case "model":
-		role = "assistant"
-	case "tool":
-		role = "user"
-	}
-
-	blocks := make([]contentBlock, 0, len(h.Parts)+len(h.TransientParts))
-	// Process standard parts
-	for _, p := range h.Parts {
-		block, ok, err := c.partToContentBlock(p, role)
-		if err != nil {
-			return "", nil, err
-		}
-		if ok {
-			blocks = append(blocks, block)
-		}
-	}
-
-	// Process transient parts
-	for _, p := range h.TransientParts {
-		block, ok, err := c.partToContentBlock(p, role)
-		if err != nil {
-			return "", nil, err
-		}
-		if ok {
-			blocks = append(blocks, block)
-		}
-	}
-	return role, blocks, nil
-}
-
-func (c *client) partToContentBlock(p *llm.Part, role string) (contentBlock, bool, error) {
-	if p.FunctionCall != nil {
-		// Fail fast if tool call has an empty ID - Anthropic requires ID for tool_use
-		if p.FunctionCall.ID == "" {
-			c.logger.Error("Encountered tool call with empty ID during serialization", "tool_name", p.FunctionCall.Name)
-			return contentBlock{}, false, fmt.Errorf("invalid tool payload: missing ID for tool call '%s'", p.FunctionCall.Name)
-		}
-		args := p.FunctionCall.Args
-		if args == nil {
-			args = make(map[string]interface{})
-		}
-		argsJSON, _ := json.Marshal(args)
-		return contentBlock{
-			Type:  "tool_use",
-			ID:    p.FunctionCall.ID,
-			Name:  p.FunctionCall.Name,
-			Input: json.RawMessage(argsJSON),
-		}, true, nil
-	}
-	if p.FunctionResponse != nil {
-		// Fail fast if tool response has an empty ID - Anthropic requires tool_use_id
-		if p.FunctionResponse.ID == "" {
-			c.logger.Error("Encountered tool response with empty ID during serialization", "tool_name", p.FunctionResponse.Name)
-			return contentBlock{}, false, fmt.Errorf("invalid tool payload: missing ID for tool response '%s'", p.FunctionResponse.Name)
-		}
-		return contentBlock{
-			Type:      "tool_result",
-			ToolUseID: p.FunctionResponse.ID,
-			Content:   marshalResponse(p.FunctionResponse.Response),
-		}, true, nil
-	}
-	if p.IsThought && role == "assistant" {
-		return contentBlock{
-			Type:      "thinking",
-			Thinking:  p.Text,
-			Signature: string(p.ThoughtSignature),
-		}, true, nil
-	}
-	if p.Text != "" {
-		return contentBlock{
-			Type: "text",
-			Text: p.Text,
-		}, true, nil
-	}
-	return contentBlock{}, false, nil
-}
-
-func (c *client) appendOrMergeMessage(messages []message, role string, blocks []contentBlock) []message {
-	if len(messages) > 0 && messages[len(messages)-1].Role == role {
-		messages[len(messages)-1].Content = append(messages[len(messages)-1].Content, blocks...)
-		return messages
-	}
-	return append(messages, message{
-		Role:    role,
-		Content: blocks,
-	})
-}
-
-func (c *client) toAnthropicTools(decls []*tools.ToolDeclaration) []tool {
-	if len(decls) == 0 {
-		return nil
-	}
-	res := make([]tool, 0, len(decls))
-	for _, d := range decls {
-		schema := toAnthropicSchema(d.Parameters)
-		if schema == nil {
-			// Anthropic requires a valid object schema even for parameterless tools
-			schema = map[string]interface{}{
-				"type":       "object",
-				"properties": map[string]interface{}{},
-			}
-		}
-		res = append(res, tool{
-			Name:        d.Name,
-			Description: d.Description,
-			InputSchema: schema,
-		})
-	}
-	return res
-}
-
-func toAnthropicSchema(s *tools.Schema) interface{} {
-	if s == nil {
-		return nil
-	}
-	res := map[string]interface{}{
-		"type": strings.ToLower(s.Type),
-	}
-	if s.Description != "" {
-		res["description"] = s.Description
-	}
-	if len(s.Enum) > 0 {
-		res["enum"] = s.Enum
-	}
-
-	// Only add properties if there are entries
-	if len(s.Properties) > 0 {
-		props := make(map[string]interface{})
-		for k, v := range s.Properties {
-			props[k] = toAnthropicSchema(v)
-		}
-		res["properties"] = props
-	}
-
-	// Only add required if there are entries
-	if len(s.Required) > 0 {
-		res["required"] = s.Required
-	}
-
-	// Only add items for arrays
-	if strings.ToLower(s.Type) == "array" && s.Items != nil {
-		res["items"] = toAnthropicSchema(s.Items)
-	}
-
-	return res
-}
-
-func (c *client) fromAnthropicResponse(resp *messagesResponse, duration float64) (*llm.Content, *llm.Metrics, error) {
-	content := c.extractContent(resp)
-	metrics := c.buildMetrics(resp, duration)
-
-	// Detect output-budget truncation. Anthropic returns
-	// stop_reason="max_tokens" when the response was cut off because
-	// the request's max_tokens cap was reached. For tool_use blocks,
-	// this means the JSON `input` object was emitted only partially —
-	// the outer braces are well-formed (Anthropic closes them) so
-	// json.Unmarshal succeeds, but keys that hadn't been emitted yet
-	// (often the largest, last-emitted key like `content`) are missing.
-	//
-	// Surfacing this as an error lets the caller decide whether to
-	// retry with a larger MaxTokens budget (via WithMaxTokens) or to
-	// break the call into smaller pieces. The alternative — silently
-	// passing the truncated content downstream — caused the registry
-	// to reject calls with cryptic "missing required parameters"
-	// errors that the model would then "retry" identically, looping
-	// indefinitely.
-	//
-	// The check is universal (not tool_use-specific): even text-only
-	// responses can be cut mid-thought in ways that mislead the agent
-	// into acting on incomplete reasoning. Conservative policy: any
-	// truncation is an error; the caller decides recovery.
-	//
-	// Pinned by TestStopReasonMaxTokens_ProducesError,
-	// TestStopReasonMaxTokens_TextOnlyStillReturnsError,
-	// TestStopReasonEndTurn_NoError, and TestStopReasonToolUse_NoError
-	// in truncation_test.go.
-	if err := checkTruncation(resp); err != nil {
-		return content, metrics, err
-	}
-
-	return content, metrics, nil
-}
-
-// checkTruncation reports whether the response was cut off by the
-// output-token budget. See fromAnthropicResponse for the rationale.
-//
-// The error message intentionally avoids any substring that
-// llmerr.Classify treats as transient (HTTP status patterns, "rate
-// limit", "503", etc.) so the resilient client falls into Classify's
-// terminal default branch and does not auto-retry. Pinned by
-// TestTruncationError_IsTerminal.
-func checkTruncation(resp *messagesResponse) error {
-	if resp == nil || resp.StopReason != "max_tokens" {
-		return nil
-	}
-
-	// Provide a tool_use-aware diagnostic when we can identify it as
-	// the truncation site, since that is the most common and most
-	// damaging case.
-	for _, block := range resp.Content {
-		if block.Type == "tool_use" {
-			return fmt.Errorf(
-				"response truncated at max_tokens during tool_use "+
-					"(tool=%q): the tool arguments are incomplete and "+
-					"cannot be safely dispatched. Increase MaxTokens "+
-					"via WithMaxTokens, or break the tool call into "+
-					"smaller pieces",
-				block.Name,
-			)
-		}
-	}
-	return fmt.Errorf("response truncated at max_tokens: output budget " +
-		"was exhausted before the model finished. Increase MaxTokens " +
-		"via WithMaxTokens, or shorten the prompt/response")
-}
-
-// extractContent deserializes response content blocks into domain llm.Part objects.
-func (c *client) extractContent(resp *messagesResponse) *llm.Content {
-	content := &llm.Content{
-		Role:  "model",
-		Parts: make([]*llm.Part, 0, len(resp.Content)),
-	}
-
-	for _, block := range resp.Content {
-		switch block.Type {
-		case "text":
-			if block.Text != "" {
-				content.Parts = append(content.Parts, &llm.Part{Text: block.Text})
-			}
-		case "thinking":
-			content.Parts = append(content.Parts, &llm.Part{
-				Text:             block.Thinking,
-				IsThought:        true,
-				ThoughtSignature: []byte(block.Signature),
-			})
-		case "tool_use":
-			var args map[string]interface{}
-			if m, ok := block.Input.(map[string]interface{}); ok {
-				args = m
-			}
-			content.Parts = append(content.Parts, &llm.Part{
-				FunctionCall: &llm.FunctionCall{
-					ID:   block.ID,
-					Name: block.Name,
-					Args: args,
-				},
-			})
-		}
-	}
-
-	content.Validate() // Final boundary sanitization
-	return content
-}
-
-// buildMetrics assembles llm.Metrics with Vertex AI-specific token accounting and traffic type resolution.
-func (c *client) buildMetrics(resp *messagesResponse, duration float64) *llm.Metrics {
-	promptTokens := resp.Usage.InputTokens
-	if c.isVertex() {
-		// Vertex AI reports input_tokens as only the newly added tokens (delta).
-		// Total context = input_tokens + cache_creation + cache_read
-		promptTokens += resp.Usage.CacheReadInputTokens + resp.Usage.CacheCreationInputTokens
-	}
-
-	metrics := &llm.Metrics{
-		Model:          c.model,
-		PromptTokens:   promptTokens,
-		ResponseTokens: resp.Usage.OutputTokens,
-		// Anthropic does not separately report reasoning tokens; the
-		// count is rolled into OutputTokens. Setting this to 0 ensures
-		// the pricing layer's
-		//   OutputCost = ResponseTokens × Comp + ThinkingTokens × Thinking
-		// reduces to ResponseTokens × Comp, which is the wire-correct
-		// charge. See issue #72 and ADR-023.
-		ThinkingTokens:   0,
-		CachedTokens:     resp.Usage.CacheReadInputTokens,
-		CacheWriteTokens: resp.Usage.CacheCreationInputTokens,
-		TotalTokens:      promptTokens + resp.Usage.OutputTokens,
-		Duration:         duration,
-		TrafficType:      c.resolveTrafficType(resp),
-	}
-
-	// Log token throughput for diagnostics
-	if metrics.ResponseTokens > 0 && metrics.Duration > 0.1 {
-		tokensPerSec := float64(metrics.ResponseTokens) / metrics.Duration
-		c.logger.Debug("token_throughput",
-			"platform", runtime.GOOS,
-			"provider", "anthropic",
-			"model", c.model,
-			"response_tokens", metrics.ResponseTokens,
-			"duration_sec", metrics.Duration,
-			"tokens_per_sec", tokensPerSec,
-			"cached_tokens", metrics.CachedTokens,
-			"cache_creation_tokens", resp.Usage.CacheCreationInputTokens,
-		)
-	}
-
-	return metrics
-}
-
-// resolveTrafficType determines the traffic type via a two-tier fallback:
-// 1. Primary: server response metadata (extra_properties.google.traffic_type)
-// 2. Secondary: reflected intent from the x-vertex-ai-llm-shared-request-type header
-func (c *client) resolveTrafficType(resp *messagesResponse) string {
-	// 1. Primary: Source of Truth (Server Response Metadata)
-	if resp.Usage.ExtraProperties != nil && resp.Usage.ExtraProperties.Google != nil && resp.Usage.ExtraProperties.Google.TrafficType != "" {
-		return resp.Usage.ExtraProperties.Google.TrafficType
-	}
-
-	// 2. Secondary: Fallback (Reflected Intent from Headers)
-	for k, v := range c.headers {
-		normalizedK := strings.ReplaceAll(strings.ToLower(k), "_", "-")
-		if normalizedK == "x-vertex-ai-llm-shared-request-type" && strings.TrimSpace(strings.ToLower(v)) == "priority" {
-			return "ON_DEMAND_PRIORITY"
-		}
-	}
-
-	return ""
-}
-
 func (c *client) GenerateImages(ctx context.Context, model, prompt string, mimeType string) ([][]byte, error) {
 	return nil, fmt.Errorf("GenerateImages not implemented for Anthropic")
 }
@@ -777,14 +406,3 @@ func (c *client) buildHTTPRequest(ctx context.Context, body []byte) (*http.Reque
 	return req, nil
 }
 
-func marshalResponse(res map[string]interface{}) string {
-	if res == nil {
-		return ""
-	}
-	// Typically we want the 'result' field if it exists, otherwise the whole thing
-	if val, ok := res["result"].(string); ok {
-		return val
-	}
-	b, _ := json.Marshal(res)
-	return string(b)
-}
